requirements.txt
click==8.1.7
dlib==19.24.2
face-recognition
face-recognition-models==0.3.0
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
python-dotenv==1.0.0
